requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
dill==0.4.0
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
graphframes==0.6
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.63
langchain-openai==0.3.19
langchain-text-splitters==0.3.8
langsmith==0.3.44
Levenshtein==0.27.1
modAL-python==0.4.2.1
multipledispatch==1.0.0
neo4j==5.28.1
nose==1.3.7
numpy==2.0.2
openai==1.84.0
openpyxl==3.1.5
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pluggy==1.6.0
pyarrow==20.0.0
pydantic==2.11.5
pydantic_core==2.33.2
pyspark==3.5.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-Levenshtein==0.27.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
skorch==0.9.0
sniffio==1.3.1
Spark-Matcher==0.3.2
SQLAlchemy==2.0.41
tabulate==0.9.0
tenacity==9.1.2
thefuzz==0.22.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
zstandard==0.23.0
